requirements.txt
flask
requests
openai
google-cloud-texttospeech
google-cloud-speech
python-dotenv
gunicorn
